add subject festures

+++ backend/src/controllers/teacher.controller.js
@@ -5,10 +5,10 @@ import jwt from "jsonwebtoken";
 
 const registerTeacher = asyncHandler(async (req, res) => {
     const { name, email, gender, address, phoneNumber, password } = req.body;
+
     if (!name || !email || !gender || !address || !phoneNumber || !password) {
         return res.status(400).json(new ApiResponse(400, "Please provide all required fields"));
     }
-
     const existingTeacher = await Teacher.findOne({ email });
     if (existingTeacher) {
         return res.status(409).json(new ApiResponse(409, "Email already in use"));
@@ -54,10 +54,10 @@ const loginTeacher = asyncHandler(async (req, res) => {
 });
 
 const getAllTeachers = asyncHandler(async (req, res) => {
-    const teachers = await Teacher.find();
+    const teachers = await Teacher.find().select("-password");
 
     if (!teachers || teachers.length === 0) {
-        return res.status(404).json(new ApiResponse(404, "No teachers found"));
+        return res.status(404).json(new ApiResponse(404, "No teachers found", []));
     }
 
     return res.status(200).json(new ApiResponse(200, "Teachers retrieved successfully", { teachers }));
@@ -65,6 +65,9 @@ const getAllTeachers = asyncHandler(async (req, res) => {
 
 const deleteTeacher = asyncHandler(async (req, res) => {
     const { id } = req.params;
+    if (!id) {
+        return res.status(404).json(new ApiResponse(404, "Please provide teacher id"));
+    }
 
     const teacher = await Teacher.findByIdAndDelete(id);
